db: move disk usage code to separate file
